# clean_website/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Study Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /materials/{id}/summaries:
    get:
      operationId: listSummaries
      tags: [summaries]
      summary: List summaries for a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Summary"
    post:
      operationId: generateSummary
      tags: [summaries]
      summary: Generate a summary for a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SummaryRequest"
      responses:
        "201":
          description: Generated summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"
  /summaries/{id}:
    get:
      operationId: getSummary
      tags: [summaries]
      summary: Get a summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"
    delete:
      operationId: deleteSummary
      tags: [summaries]
      summary: Delete a summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Flashcards ────────────────────────────────────────────────────────────
...
  /materials/{id}/question-sets:
    get:
      operationId: listQuestionSets
      tags: [questions]
      summary: List question sets for a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question sets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuestionSet"
    post:
      operationId: generateQuestions
      tags: [questions]
      summary: Generate questions from material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionRequest"
      responses:
        "201":
          description: Generated question set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionSet"
  /question-sets/{id}:
    get:
      operationId: getQuestionSet
      tags: [questions]
      summary: Get a question set with questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionSet"
    delete:
      operationId: deleteQuestionSet
      tags: [questions]
      summary: Delete a question set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Exams ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Summary:
      type: object
      required: [id, materialId, summaryType, language, content, createdAt]
      properties:
        id:
          type: integer
        materialId:
          type: integer
        summaryType:
          type: string
          enum: [quick, detailed, chapter, topic, key_takeaways, exam_focused]
        language:
          type: string
          enum: [he, en]
        content:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    SummaryRequest:
      type: object
      required: [summaryType, language]
      properties:
        summaryType:
          type: string
          enum: [quick, detailed, chapter, topic, key_takeaways, exam_focused]
        language:
          type: string
          enum: [he, en]
        topic:
          type: string

    # ── Flashcards ────────────────────────────────────────────────────────────
...
    QuestionSet:
      type: object
      required: [id, materialId, title, language, createdAt]
      properties:
        id:
          type: integer
        materialId:
          type: integer
        title:
          type: string
        language:
          type: string
          enum: [he, en]
        questionCount:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        createdAt:
          type: string
          format: date-time
    Question:
      type: object
      required: [id, setId, questionType, question, answer, createdAt]
      properties:
        id:
          type: integer
        setId:
          type: integer
        questionType:
          type: string
          enum: [open, multiple_choice, true_false, fill_blank, short_answer]
        question:
          type: string
        answer:
          type: string
        explanation:
          type: ["string", "null"]
        options:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        createdAt:
          type: string
          format: date-time
    QuestionRequest:
      type: object
      required: [language]
      properties:
        language:
          type: string
          enum: [he, en]
        questionCount:
          type: integer
        questionTypes:
          type: array
          items:
            type: string
            enum: [open, multiple_choice, true_false, fill_blank, short_answer]
        difficulty:
          type: string
          enum: [easy, medium, hard, mixed]

    # ── Exams ─────────────────────────────────────────────────────────────────
...